Make code compatible with both 4 and 5 versions of ESP-IDF

See merge request particle-sensor/internal-esp-idf!21

+++ main/AppMain.cpp
@@ -186,7 +186,7 @@ void loop()
             break;
         case DustMonitorController::ProcessStatus::NeedRefreshClock:
             {
-                const auto delay = std::min((uint32_t)getMicrosecondsTillNextMinute() / 1000 + 1, 100u);
+                const auto delay = std::min(getMicrosecondsTillNextMinute() / 1000 + 1, 100ll);
                 DEBUG_LOG("Refresh clock in " << delay << " ms")
                 embedded::delay(delay);
             }
